+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
-spring.mail.username=
-spring.mail.password=
+spring.mail.username=mirceanghel98@gmail.com
+spring.mail.password=ijya yczz mogz lwwy
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
