+++ src/main/resources/templates/admin_page.html
@@ -13,6 +13,11 @@
         <p th:text="${user.firstName}"></p>
         <p th:text="${user.lastName}"></p>
         <p th:text="${user.mail}"></p>
+        <button type="button" th:onclick="@{/api/users/delete/${user.id}}" th:object="${user}">X</button>
+<!--        <button type="button" onclick="location.href='/api/delete/'"<h4>&nbsp;&nbsp;Delete User&nbsp;&nbsp;</h4></button>-->
+<!--        <button type="button" th:action="@{'/api/users/delete/{id}'(id=${user.id})}" th:method="delete"><h4>&nbsp;&nbsp;Delete User&nbsp;&nbsp;</h4></button>-->
+<!--        <a th:href="|@{/api/users/delete/${user.id}}|">Delete</a>-->
+<!--        <button type="button" th:action="/api/users/{id}"><h4>&nbsp;&nbsp;Delete User&nbsp;&nbsp;</h4></button>-->
         <p><hr></hr></p>
     </li>
 </ul>
