+++ src/main/resources/templates/error_page.html
@@ -8,7 +8,7 @@
 
 <h1 align="center">Something Went Wrong! Invalid Input!</h1>
 
-<button type="button" onclick="location.href='/api/mainpage'"><h4>&nbsp;&nbsp;Go home&nbsp;&nbsp;</h4></button>
+<button type="button" onclick="location.href='/api/mainpage'"><h4>&nbsp;&nbsp;Go Home&nbsp;&nbsp;</h4></button>
 
 </body>
 </html>
