+++ src/main/resources/templates/user_form.html
@@ -7,6 +7,7 @@
 <body bgcolor="#5f9ea0">
 
 <h2 align="center">Form Page</h2>
+<button type="button" onclick="location.href='/api/mainpage'"><h4>&nbsp;&nbsp;Go Home&nbsp;&nbsp;</h4></button>
 
 <form action="#" th:action="@{/api/form/submit}" th:object="${userDTO}" method="post">
     <p>First Name <input type="text" th:field="*{firstName}"/></p>
