+++ src/main/resources/templates/user_page.html
@@ -6,7 +6,7 @@
 </head>
 <body bgcolor="#5f9ea0">
 <h2 align="center">Account Details</h2>
-<button type="button" onclick="location.href='/api/mainpage'"><h4>&nbsp;&nbsp;Go home&nbsp;&nbsp;</h4></button>
+<button type="button" onclick="location.href='/api/mainpage'"><h4>&nbsp;&nbsp;Go Home&nbsp;&nbsp;</h4></button>
 
 <p th:text="${user.id}"></p>
 <p th:text="${user.firstName}"></p>
